Rplot Friends Week1
Disclaimer: I changed my corpus. This description of my corpus differs from the one submitted on Wednesday.

Moving to Amsterdam made me realize, that it is really difficult to share the same taste in music with my expat friends here in Amsterdam. Growing up in Germany, my friends and I were used to different kinds of music but growing up together made us associate periods of our lives with music that makes us feel connected even years later. Unfortunately, with the friends that I have in Amsterdam, the only songs that we can sing along are either songs from the past (ABBA, Linkin Park) or songs played on the radio (Taylor Swift, the Weekend). But none of us is listening to that kind of music in private. So I have created a corpus consisting of the recap Spotify playlists from last year of my close friends in Amsterdam, to investigate which direction of music is the underlying connection in music taste we share. So far I have created a plot showing the difference in energy, valence and loudness of my three friends (Lorah, Io, and Anett). So far I can only say that Io's music is the highest in energy and loudness compared to the other two.